JK: Improved form handling and the generation of the query string.     Separation of the parameters from the URL name when following links     and when submitting forms.
*** ATTENTION FOR JAVA PEOPLE***

When the separation is done, the parameters are
given to the GetObjectWWW function inside the form_data parameter.
This parameter shouldn't be freed by the function. If the mode has the
AMAYA_FORM_PAST flag, you should send the form_data in a post request;
otherwise, you should concatenate it to the URL using the '?' char and
send it using a GET.