Bug fix
When an image generates a presentation box which is also an image,
Thot is confused, it mixes the data about the two images and crashes
when releasing the images.

This occurs in amaya when an IMG has an attribute id. The id attribute
generates a taget icon, which is also an image.

Vincent.