Localization stuff for Amaya on Windoz.
1. a resource per language. There is 3 languages used to be maintained at the present time:
- French
- English
- German

The convetion is to use the prefix FR_ for French dialogs, EN_ for English Dialogs
and DE_ for German Dialogs. For example if one has a resource identified by
OPENDOC, on should have:

FR_OPENDOC
EN_OPENDOC
DE_OPENDOC

Of course there is only one callback for each resource.

Ramzi.